+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.0
+
+* Refactored the error log
+* added some utilities
+* added the ability to choose between 'local' and 'session' storage
+* added expiration option when setting the storage 
+
 ## 1.0.1
 
 * Minor changes to the README file.